+++ rye/src/platform.rs
@@ -5,6 +5,7 @@ use std::{env, fs};
 
 use anyhow::{anyhow, Context, Error};
 
+use crate::config::Config;
 use crate::pyproject::latest_available_python_version;
 use crate::sources::{PythonVersion, PythonVersionRequest};
 
@@ -169,28 +170,27 @@ pub fn list_known_toolchains() -> Result<Vec<(PythonVersion, PathBuf)>, Error> {
     Ok(rv)
 }
 
-/// Returns the default author from git.
-pub fn get_default_author() -> Option<(String, String)> {
-    let rv = Command::new("git")
+/// Returns the default author from git or the config.
+pub fn get_default_author_with_fallback() -> Option<(String, String)> {
+    let (mut name, mut email) = Config::current().default_author();
+
+    if let Ok(rv) = Command::new("git")
         .arg("config")
         .arg("--get-regexp")
         .arg("^user.(name|email)$")
         .stdout(Stdio::piped())
         .output()
-        .ok()?;
-
-    let mut name = None;
-    let mut email = None;
-
-    for line in std::str::from_utf8(&rv.stdout).ok()?.lines() {
-        match line.split_once(' ') {
-            Some((key, value)) if key == "user.email" => {
-                email = Some(value.to_string());
+    {
+        for line in std::str::from_utf8(&rv.stdout).ok()?.lines() {
+            match line.split_once(' ') {
+                Some((key, value)) if key == "user.email" && email.is_none() => {
+                    email = Some(value.to_string());
+                }
+                Some((key, value)) if key == "user.name" && name.is_none() => {
+                    name = Some(value.to_string());
+                }
+                _ => {}
             }
-            Some((key, value)) if key == "user.name" => {
-                name = Some(value.to_string());
-            }
-            _ => {}
         }
     }